Fix lib

+++ Userland/c/usrlib.c
@@ -149,8 +149,8 @@ void changeToKeys() {
 
 char* concat(char *s1, char *s2)
 {
-    size_t len1 = strlen(s1);
-    size_t len2 = strlen(s2);
+    int64_t len1 = strlen(s1);
+    int64_t len2 = strlen(s2);
     char *result = malloc(len1+len2+1);//+1 for the zero-terminator
     //in real code you would check for errors in malloc here
     memcpy(result, s1, len1);
